Reintroduce MACos, checkout install and don't use curl
After some investigation on the CI runner, a likely potential reason that macos CI jobs fail is that GitHub will rate-limit curl scripts if the user is unauthenticated which the CI clusters are. As eventually we use the script in `lf-lang/installation` it might be a better idea to not use curl and get rate-limited but use checkout directly. This also reduces the dependency of `lf-lang.org` behaving correctly.

MacOS CI is also reintroduced in this commit.

Signed-off-by: axmmisaka <axmmisaka@users.noreply.github.com>